add guestIds to Room class and change the HotelManagementApp
